remove test page
